Rename and move Validation error model
