+++ api/consumer/readDocumentReference/tests/unit/test_index.py
@@ -1,29 +1,8 @@
 import os
 from unittest import mock
-from pydantic import BaseModel, ValidationError, validator
 import pytest
-from typing import Any
 
-from lambda_utils.tests.unit.utils import make_aws_event
-from lambda_pipeline.types import PipelineData, LambdaContext, FrozenDict
-from aws_lambda_powertools.utilities.parser.models import APIGatewayProxyEventModel
-
-
-class DummyModel(BaseModel):
-    foo: bool
-
-    @validator("foo")
-    def something(value):
-        raise ValueError
-
-
-def handler_four_hundred(
-    data: PipelineData,
-    context: LambdaContext,
-    event: APIGatewayProxyEventModel,
-    dependencies: FrozenDict[str, Any],
-) -> PipelineData:
-    DummyModel(foo="1")
+from lambda_utils.tests.unit.utils import make_aws_event, handler_four_hundred
 
 
 @mock.patch.dict(os.environ, {"AWS_REGION": "eu-west-2"}, clear=True)
